Moves files out of analyse package and puts them all in resource package, updates README
