+++ searcher/resource/resource_test.go
@@ -9,13 +9,16 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+const mockArticlePath = "./mock_article.json"
+
 var (
 	// MockArticle is used for testing packages
 	mockArticle = &Resource{}
+	result      []string
 )
 
 func init() {
-	data, err := ioutil.ReadFile("../mocks/mock_article.json")
+	data, err := ioutil.ReadFile(mockArticlePath)
 	if err != nil {
 		log.Info().Err(err).Msg("reading from mock article file")
 	}
@@ -23,6 +26,34 @@ func init() {
 	if err := json.Unmarshal(data, mockArticle); err != nil {
 		log.Info().Err(err).Msg("unmarshalling JSON")
 	}
+}
+
+func Test_WordCount(t *testing.T) {
+	expectedWordCount := 9
+	content := []string{"python is the best!", "Especially Django - I love it."}
+
+	assert.Equal(t, expectedWordCount, WordCount(content))
+}
+
+func Test_FindTopics(t *testing.T) {
+	expectedTopics := []string{"python", "Django"}
+	content := []string{"python is the best!", "Especially Django - I love it."}
+
+	assert.Equal(t, expectedTopics, FindTopics(content))
+}
+
+func Benchmark_FindTopics(b *testing.B) {
+	var topics []string
+	content := mockArticle.document.content
+	// run the FindTopics function b.N times
+	for n := 0; n < b.N; n++ {
+		topics = FindTopics(content)
+	}
+
+	// always store the result to a package level variable
+	// so the compiler cannot eliminate the Benchmark itself.
+	// https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go
+	result = topics
 }
 
 func Test_Analyse(t *testing.T) {
